hyperv/stor: Support UNMAP/TRIM for FreeBSD on Win2012R2
UNMAP/TRIM is not supported if SCSI version is below SPC-3.
FreeBSD on Win2012R2 gets SCSI version SPC-2, as a result, UNMAP
is not supported. So, here we check it and set SPC-3 instead because
the host supported SPC-3 for Msft disk.

This issue does not occur on windows 2016.

https://bugs.freebsd.org/bugzilla/show_bug.cgi?id=211549

Submitted by:	Hongjinag Zhang <honzhan microsoft com>